requirements.txt
psycopg2-binary
numpy
pandas
matplotlib
plotly-calplot
dash
dash_bootstrap_components